bugs: the Operator_querier class can not fetch operator type correctly! 1. reorganized the structure of parser, Expression and Operators 2. the classes are ready to be expanded if the bugs can be fixed

+++ src/test/test_parser.cpp
@@ -2,7 +2,7 @@
 #include "../Expression_parser.hpp"
 
 int main() {
-  expression_parser<int> parser;
+  expression_parser<double> parser;
   int t = parser.parse("(1+5) *(2+1) /3").eval();
   std::cout << t;
 }
